Downscale the image while zooming
While performing a live magnify gesture on the scroll view, swap out the
image with a downscaled version. This prevents hitches that have a
tendency to slow down WindowServer and cause the entire system to stall
for whatever reason.

I'm not sure why the scroll view doesn't just scale what the user is
able to see onscreen. I tried to override the isOpaque properties of the
views to try to activate some optimization behaviors (maybe "responsive
scrolling" can help?), but nothing seemed to work. It seems that the
image is being resized on the fly as the gesture is being done, which
can be incredibly expensive if the image is large enough (in the tens of
megabytes). This causes the hitches.

Preview.app seems to do something similar when zooming.