Counteract magnification when drawing crop target
Since the crop target is drawn inside of the scroll view, zooming in and
out naturally affects the frame of the crop target, and therefore how
thick its lines are.

Counteract this by observing the scroll view's magnification and
adjusting line thickness accordingly. This way, the borders of the crop
target appear consistent regardless of magnification level.