Only allocate cropped thumbnail backing store once
Draw directly into the thumbnail NSImage instead of using a dynamically
drawn image. This means that we only have to allocate a single bitmap
backing store. Otherwise, one seems to be allocated every time the image
is drawn. I am not sure why this is, because it should be cached by
AppKit, but alas.

We should not be creating an NSImage every time we need to display a
subsection of the original image (because it's quite expensive, and can
be invoked rapidly), but this will do for now.